feat(api): mount v1 routes
