+++ daemon/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod app;
+pub mod api;
 pub mod docker;
 pub mod errors;
 pub mod routes;
