+++ daemon/tests/apps_delete.rs
@@ -3,7 +3,7 @@ use axum::{
     http::{Request, StatusCode},
 };
 use lightshuttle_core::{
-    app::build_router,
+    api::routes::router,
     docker::{create_and_run_container, remove_container, ContainerConfig},
 };
 use tower::ServiceExt;
@@ -31,12 +31,12 @@ async fn delete_existing_app_should_succeed() {
 
     create_and_run_container(config).expect("Failed to launch container");
 
-    let app = build_router();
+    let app = router();
     let response = app
         .oneshot(
             Request::builder()
                 .method("DELETE")
-                .uri(format!("/apps/{container_name}"))
+                .uri(format!("/api/v1/apps/{container_name}"))
                 .body(Body::empty())
                 .unwrap(),
         )
@@ -56,12 +56,12 @@ async fn delete_non_existing_app_should_return_404() {
         return;
     }
 
-    let app = build_router();
+    let app = router();
     let response = app
         .oneshot(
             Request::builder()
                 .method("DELETE")
-                .uri("/apps/i-do-not-exist")
+                .uri("/api/v1/apps/i-do-not-exist")
                 .body(Body::empty())
                 .unwrap(),
         )
