+++ daemon/tests/utils.rs
@@ -1,6 +1,6 @@
 use axum::Router;
-use lightshuttle_core::app::build_router;
+use lightshuttle_core::api::routes::router;
 
 pub fn test_app() -> Router {
-    build_router()
+    router()
 }
